Config

Description

The module exposes a datatype for a Configuration file. The file must be in JSON format. If the file does not exist the module will cause a runtime exception and crash the program. Thus be sure that :set args is set to prod or dev, and the json file exists in the system path .srcconfig/cofnig.json for dev.

Documentation

data Configuration

Configuration data type

Constructors

Configuration

Fields

dbName :: Text

dbUrl :: String

dbUser :: String

dbPassword :: String

configuration :: IO (Maybe Configuration)