MongoUtils

Description

This module exposes methods for creating a mongo connection pool, and for running actions against the db.

Documentation

localPipe :: IO Pipe

Mongo must run

getPipe :: Configuration -> IO Pipe

General pipe, can connect to local instance or to a cluster db

getPool :: Configuration -> IO (Pool Pipe)

Connection pool

run :: Action IO a -> MotoTransformer a

Run a querry against the db

fromDoc :: FromBSON a => Document -> Maybe a

helper to convert from DOC to appropriate Type