Rest

Description

This module exposes a method for starting the rest endpoint

Documentation

apiRoot :: String

route for base endpoint

withJsonData :: FromJSON p => (p -> MotoTransformer a) -> ActionTransformer a

Deserialize data from json string "{object}" to the appropriate data type

withStringParam :: Text -> (String -> MotoTransformer a) -> ActionTransformer a

Deserialize data from just param to the appropriate dataType